Documentation, comments and structure
A complete readme.md, create the docker, docker-compose. And restructure python files.
