+++ .gitignore
@@ -1,0 +1,20 @@
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+
+# C Extensions
+*.so
+
+# Virtualenv
+.env/
+venv/
+ENV/
+
+# Jupyter Notebook checkpoints
+.ipynb_checkpoints
+
+# VS Code config
+.vscode/
+
+# Docker related files
+*.log
